requirements.txt
Django==1.10.3
django-cors-headers==2.0.2
django-dirtyfields==1.2.1
django-mptt==0.8.7
django-nose==1.4.4
django-qsstats-magic==0.7.2
django-tellme==0.6.3
django-autocomplete-light
patterns==0.3
Pillow==4.1.1
psycopg2==2.7.1
pycparser==2.17
pycryptodomex==3.4.6
pyparsing==2.2.0
python-dateutil==2.1
pytz==2017.2
raven==6.0.0
requests==2.6.0
six==1.10.0
django-configurations
django-crispy-forms
django-widget-tweaks
datetime